Cutting ui.py into pieces, created cam op functions in buttons_panel
